Move MainTab out of ui.py
